% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary-data.R
\docType{data}
\name{boundary}
\alias{boundary}
\title{Boundary polygon for field 10 at Bjertorp farm in southwest Sweden}
\format{A SpatialPolygonsDataFrame with the boundary polygon. The polygon has
been shrunk (i.e. the field is somewhat larger than the polygon)
Projected coordinate system Sweref99TM (epsg: 3006).}
\usage{
data(boundary)
}
\description{
Boundary polygon for field 10 at Bjertorp farm in southwest Sweden
}
\references{
Piikki, K., Söderström, M., & Stenberg, B. (2013). Sensor data
fusion for topsoil clay mapping. Geoderma, 199, 106-116.
}
\keyword{datasets}
