% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_prepare.R
\name{prepare}
\alias{prepare}
\title{Prepares Data}
\usage{
prepare(m, method, mean.filter, buff)
}
\arguments{
\item{m}{a list of all objects to be used.}

\item{method}{sampling alogorithm: 'dir' = directed (sr algorithm),
'stratdir' = stratified directed (sr algorithm), 'grid' = regular grid and
'stratrand' = random stratified.}

\item{mean.filter}{optional. side of the square window (number of
raster cells) used for mean filtering of the raster(s). default = 3}

\item{buff}{buffer zone (m) inside the sample area border, where sampling is
prohibited. default = 0.}
}
\description{
Prepares data to be used for sampling
}
